Move converter tests inline into their respective modules
Dissolve the standalone tests.rs and config_tests.rs files, placing
each test in the #[cfg(test)] mod tests block of the module it
exercises: config serde + domain matching → config.rs, pool dispatch →
worker_pool.rs, font classification → fonts.rs, URL generation →
rendering.rs, InnerVlConverter + canvas polyfill → inner.rs, and
VlConverter integration/security tests → mod.rs.
